apps/arcs/omega-monitor/requirements.txt
flask==3.0.0
gunicorn==22.0.0
nats-py==2.7.2